requirements.txt
azure-functions
requests
stressypy
aiohttp-requests==0.1.2